Make the timer precise to 100ms.
git-svn-id: http://hugogu/svn/main/Workspace/Projects/TechnicalResearch/WPF/ClearMine@82 abc3d51d-b5e4-b64c-aea7-7d3d034b83e7

+++ ClearMine.VM/ClearMineViewModel.cs
@@ -305,7 +305,7 @@
 
         public string Time
         {
-            get { return (game.UsedTime / 1000).ToString(); }
+            get { return ((double)game.UsedTime / 1000).ToString(); }
         }
 
         public string RemainedMines
